requirements.txt
selenium
pytest
requests
pyyaml
pytest-xdist
allure-pytest
